+++ ratarmount.py
@@ -37,6 +37,11 @@ try:
 except ImportError:
     print( "[Warning] The indexed_zstd module was not found. Please install it to open zstd compressed TAR files!" )
 
+try:
+    import lzmaffi
+except ImportError:
+    print( "[Warning] The lzmaffi module was not found. Please install it to open xz compressed TAR files!" )
+
 import fuse
 
 
@@ -308,7 +313,20 @@ class SQLiteIndexedTar:
         self.tarFileObject, self.rawFileObject, self.compression, self.isTar = \
             SQLiteIndexedTar._openCompressedFile( fileObject, gzipSeekPointSpacing, encoding )
 
-        # Determining if there are many frames in zst is O(1) with is_multiframe
+        if self.compression == 'xz':
+            try:
+                if len( self.tarFileObject.block_boundaries ) <= 1:
+                    print( "[Warning] The specified file '{}'".format( self.tarFileName ) )
+                    print( "[Warning] is compressed using xz but only contains one xz block. This makes it " )
+                    print( "[Warning] impossible to use true seeking! Please (re)compress your TAR using pixz" )
+                    print( "[Warning] (see https://github.com/vasi/pixz) in order for ratarmount to do be able " )
+                    print( "[Warning] to do fast seeking to requested files." )
+                    print( "[Warning] As it is, each file access will decompress the whole TAR from the beginning!" )
+                    print()
+            except:
+                pass
+
+        # Determining if there are many frames in zstd is O(1) with is_multiframe
         if self.compression == 'zst':
             try:
                 if not self.tarFileObject.is_multiframe():
@@ -456,8 +474,7 @@ class SQLiteIndexedTar:
         """ )
         return sqlConnection
 
-    @staticmethod
-    def _updateProgressBar( progressBar, fileobj ):
+    def _updateProgressBar( self, progressBar, fileobj ):
         try:
             if 'IndexedBzip2File' in globals() and isinstance( fileobj, IndexedBzip2File ):
                 # Note that because bz2 works on a bitstream the tell_compressed returns the offset in bits
@@ -470,6 +487,10 @@ class SQLiteIndexedTar:
 
             if hasattr( fileobj, 'fileobj' ):
                 progressBar.update( fileobj.fileobj().tell() )
+                return
+
+            if self.rawFileObject and hasattr( self.rawFileObject, 'tell' ):
+                progressBar.update( self.rawFileObject.tell() )
                 return
 
             progressBar.update( fileobj.tell() )
@@ -654,7 +675,7 @@ class SQLiteIndexedTar:
         if fileCount == 0:
             tarInfo = os.fstat( fileObject.fileno() )
             fname = os.path.basename( self.tarFileName )
-            for suffix in [ '.gz', '.bz2', '.bzip2', '.gzip', '.zst', '.zstd' ]:
+            for suffix in [ '.gz', '.bz2', '.bzip2', '.gzip', '.xz', '.zst', '.zstd' ]:
                 if fname.lower().endswith( suffix ) and len( fname ) > len( suffix ):
                     fname = fname[:-len( suffix )]
                     break
@@ -1068,6 +1089,9 @@ class SQLiteIndexedTar:
                 if compression == 'gz' and 'IndexedGzipFile' not in globals():
                     raise Exception( "Can't open a bzip2 compressed TAR file '{}' without indexed_gzip module!"
                                      .format( name ) )
+                if compression == 'xz' and 'lzmaffi' not in globals():
+                    raise Exception( "Can't open xz compressed TAR file '{}' without lzmaffi module!"
+                                     .format( name ) )
 
                 if oldOffset is not None:
                     fileobj.seek( oldOffset )
@@ -1086,6 +1110,11 @@ class SQLiteIndexedTar:
                 if compression == 'gz':
                     gzip.open( fileobj if fileobj else name ).read( 1 )
                     return isTar, compression
+
+                if compression == 'xz':
+                    lzmaffi.open( fileobj if fileobj else name ).read( 1 )
+                    return isTar, compression
+
             except:
                 if oldOffset is not None:
                     fileobj.seek( oldOffset )
@@ -1137,6 +1166,9 @@ class SQLiteIndexedTar:
             tarFile = IndexedGzipFile( fileobj = rawFile,
                                        drop_handles = False,
                                        spacing = gzipSeekPointSpacing )
+        elif compression == 'xz':
+            rawFile = tarFile # save so that garbage collector won't close it!
+            tarFile = lzmaffi.open( rawFile )
         elif compression == 'zst':
             rawFile = tarFile # save so that garbage collector won't close it!
             tarFile = IndexedZstdFile( rawFile.fileno() )
@@ -1256,6 +1288,8 @@ class SQLiteIndexedTar:
                 db.execute( 'INSERT INTO gzipindex VALUES (?)', ( file.read(), ) )
             db.commit()
             os.remove( gzindex )
+
+        # Note that for xz seeking loading and storing block indexes is unnecessary because it has an index included!
 
 
 class TarMount( fuse.Operations ):
@@ -1669,6 +1703,9 @@ class TarFileType:
                 if compression == 'gz':
                     gzip.open( tarFile ).read( 1 )
                     return ( tarFile, compression )
+                if compression == 'xz':
+                    lzmaffi.open( tarFile ).read( 1 )
+                    return ( tarFile, compression )
                 if compression == 'zst':
                     IndexedZstdFile( tarFile ).read( 1 )
                     return ( tarFile, compression )
@@ -1908,6 +1945,8 @@ version, you can simply do:
         compressions += [ 'bz2' ]
     if 'IndexedGzipFile' in globals():
         compressions += [ 'gz' ]
+    if 'lzmaffi' in globals():
+        compressions += [ 'xz' ]
     if 'IndexedZstdFile' in globals():
         compressions += [ 'zst' ]
     args.mount_source = [ TarFileType( mode = 'r', compressions = compressions, encoding = args.encoding )( tarFile )[0]
@@ -1917,7 +1956,10 @@ version, you can simply do:
     # Automatically generate a default mount path
     if args.mount_point is None:
         tarPath = args.mount_source[0]
-        for doubleExtension in [ '.tar.bz2', '.tar.gz' ]:
+        for compression in compressions:
+            if not compression:
+                continue
+            doubleExtension = '.tar.' + compression
             if tarPath[-len( doubleExtension ):].lower() == doubleExtension.lower():
                 args.mount_point = tarPath[:-len( doubleExtension )]
                 break