+++ tests/runtests.sh
@@ -562,6 +562,7 @@ for (( iTest = 0; iTest < ${#tests[@]}; iTest += 3 )); do
     toCleanUp=()
     tbz2Path=
     tgzPath=
+    txzPath=
     tzstPath=
     isSingleArchive=0
 
@@ -578,13 +579,19 @@ for (( iTest = 0; iTest < ${#tests[@]}; iTest += 3 )); do
             toCleanUp+=( "$tarPath" )
             gzip --keep --stdout --decompress -- "$tgzPath" > "$tarPath"
             ;;
+        *.tar.xz|*.txz)
+            txzPath=$tarPath
+            tarPath=$( mktemp --suffix='.tar' )
+            toCleanUp+=( "$tarPath" )
+            pixz -d "$txzPath" "$tarPath"
+            ;;
         *.tar.zst|*.tzst)
             tzstPath=$tarPath
             tarPath=$( mktemp --suffix='.tar' )
             toCleanUp+=( "$tarPath" )
             zstd --keep --stdout --decompress -- "$tzstPath" > "$tarPath"
             ;;
-        *.bz2|*.gz|*.zst)
+        *.bz2|*.gz|*.xz|*.zst)
             isSingleArchive=1
             ;;
     esac
@@ -601,6 +608,12 @@ for (( iTest = 0; iTest < ${#tests[@]}; iTest += 3 )); do
         gzip --keep --stdout "$tarPath" > "$tgzPath"
     fi
 
+    if [[ -z "$txzPath" && "$isSingleArchive" -ne 1 ]]; then
+        txzPath=$( mktemp --suffix='.tar.xz' )
+        toCleanUp+=( "$txzPath" )
+        pixz "$tarPath" "$txzPath"
+    fi
+
     if [[ -z "$tzstPath" && "$isSingleArchive" -ne 1 ]]; then
         tzstPath=$( mktemp --suffix='.tar.zst' )
         toCleanUp+=( "$tzstPath" )
@@ -610,6 +623,7 @@ for (( iTest = 0; iTest < ${#tests[@]}; iTest += 3 )); do
     checkFileInTAR "$tarPath" "$fileName" "$checksum"
     if [[ -f "$tbz2Path" ]]; then checkFileInTAR "$tbz2Path" "$fileName" "$checksum"; fi
     if [[ -f "$tgzPath" ]]; then checkFileInTAR "$tgzPath" "$fileName" "$checksum"; fi
+    if [[ -f "$txzPath" ]]; then checkFileInTAR "$txzPath" "$fileName" "$checksum"; fi
     if [[ -f "$tzstPath" ]]; then checkFileInTAR "$tzstPath" "$fileName" "$checksum"; fi
 
     for tmpFile in "${toCleanUp[@]}"; do