+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
  - Add CLI options --index-file and --index-folder to specify an index location or folder.
  - Add support for zstd compressed TARs.
+ - Add support for xz compressed TARs when compressed with pixz or pxz or similar tools
+   limiting the block size.
  - Fix index location check not accounting for SQLite special requirements.
 
 # Version 0.6.1 built on 2020-10-02